// marketplace/src/screens/MainScreen/types.ts
export enum TabValues {
  Signals = 'signals-tab',
  Favorites = 'favorites-tab',
  Popular = 'popular-tab',
  All = 'all-tab',
  Strategies = 'strategies-tab',
  Robots = 'robots-tab',
  Consultants = 'consultants-tab',
  Script = 'scripts-tab',
  Courses = 'courses-tab',
  Others = 'others-tab',
  Analystys = 'analystys-tab',
  Traders = 'traders-tab',
}